livekit-plugins/livekit-plugins-hamming/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "livekit-plugins-hamming"
dynamic = ["version"]
description = "Agent Framework Plugin for Hamming — send OTel traces, logs, and metrics to Hamming"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10.0"
authors = [{ name = "Hamming", email = "engineering@hamming.ai" }]
keywords = ["voice", "ai", "realtime", "audio", "livekit", "hamming", "observability", "opentelemetry"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "livekit-agents>=1.4.2",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[project.urls]
Documentation = "https://docs.hamming.ai"
Website = "https://hamming.ai/"
Source = "https://github.com/livekit/agents"

[tool.hatch.version]
path = "livekit/plugins/hamming/version.py"

[tool.hatch.build.targets.wheel]
packages = ["livekit"]

[tool.hatch.build.targets.sdist]
include = ["/livekit"]
